requirements.txt
Django==3.1.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.5.0
mysqlclient==2.0.1